+++ ECommerceProdutos/Application/Services/ProductService.cs
@@ -1,16 +1,19 @@
 ﻿using Application.Interfaces;
 using Domain.Entities;
 using Domain.IRepositories;
+using FluentValidation;
 
 namespace Application.Services
 {
     public class ProductService : IProductService
     {
         private readonly IProductRepository _repository;
+        private readonly IValidator<Product> _validator;
 
-        public ProductService(IProductRepository repository)
+        public ProductService(IProductRepository repository, IValidator<Product> validator)
         {
             _repository = repository;
+            _validator = validator;
         }
 
         public async Task<IEnumerable<Product>> GetAllAsync()
@@ -20,21 +23,33 @@ namespace Application.Services
 
         public async Task<Product> GetByIdAsync(Guid id)
         {
+            if (id == Guid.Empty) throw new Exception("Deve informar um id");
             return await _repository.GetByIdAsycn(id);
         }
 
         public async Task InsertAsync(Product product)
         {
+            if (product is null) throw new Exception("Deve informar um produto para inserção");
+
+            var result = await _validator.ValidateAsync(product);
+            if(!result.IsValid) throw new ValidationException(result.Errors);
+
             await _repository.InsertAsync(product);
         }
 
         public async Task RemoveAsync(Guid id)
         {
+            if (id == Guid.Empty) throw new Exception("Deve informar um id");
             await _repository.RemoveAsync(id);
         }
 
         public async Task UpdateAsync(Product product)
         {
+            if (product is null) throw new Exception("Deve informar um produto para atualização");
+
+            var result = await _validator.ValidateAsync(product);
+            if (!result.IsValid) throw new ValidationException(result.Errors);
+
             await _repository.UpdateAsync(product);
         }
     }
